Rename backgroundSize to backgroundSizes
