+++ src/utilityModules.js
@@ -4,7 +4,7 @@ import backgroundAttachment from './generators/backgroundAttachment'
 import backgroundColors from './generators/backgroundColors'
 import backgroundPosition from './generators/backgroundPosition'
 import backgroundRepeat from './generators/backgroundRepeat'
-import backgroundSize from './generators/backgroundSize'
+import backgroundSizes from './generators/backgroundSizes'
 import borderColors from './generators/borderColors'
 import borderRadius from './generators/borderRadius'
 import borderStyle from './generators/borderStyle'
@@ -51,7 +51,7 @@ export default [
   { name: 'backgroundColors', generator: backgroundColors },
   { name: 'backgroundPosition', generator: backgroundPosition },
   { name: 'backgroundRepeat', generator: backgroundRepeat },
-  { name: 'backgroundSize', generator: backgroundSize },
+  { name: 'backgroundSizes', generator: backgroundSizes },
   { name: 'borderColors', generator: borderColors },
   { name: 'borderRadius', generator: borderRadius },
   { name: 'borderStyle', generator: borderStyle },
